+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.9.0
+
+### Added
+
+- Support `textDocument.signatureHelp`
+
+### Removed
+
+- A bunch of `try/except Exception` blocks around Jedi calls. Hopefully those aren't still necessary on `jedi>=0.17`...
+
 ## 0.8.0
 
 ### Added